refactor: rename hooks (review points 4, 5)
